requirements.txt
streamlit>=1.32.0
Pillow>=9.0.0
numpy>=1.23.0
torch==2.0.1
torchvision==0.15.2